refactor: 사용자 생성 및 삭제 기능 모듈 이동
