+++ src/common/interfaces/user-payload.interface.ts
