--- src/modules/auth/dtos/jwt.dto.ts
@@ -1,4 +1,0 @@
-export class JwtPayloadDto {
-  readonly id: number;
-  readonly email: string;
-}
