+++ src/modules/auth/interfaces/authenticate-request.interface.ts
@@ -1,0 +1,5 @@
+import { UserPayload } from '../../../common/interfaces/user-payload.interface';
+
+export interface AuthenticateRequest extends Request {
+  user: UserPayload;
+}
