--- src/modules/auth/interfaces/oauth-request.interface.ts
@@ -1,5 +1,0 @@
-import { UserPayload } from './user-payload.interface';
-
-export interface OAuthRequest extends Request {
-  user: UserPayload;
-}
