+++ src/modules/auth/interfaces/reponse-login.interface.ts
@@ -1,4 +1,4 @@
-import { UserInfo } from './userInfo.inerface';
+import { UserInfo } from './user-info.interface.ts';
 
 export interface ResponseLogin {
   accessToken: string;
