+++ src/modules/auth/interfaces/user-info.interface.ts.ts
@@ -1,0 +1,7 @@
+export interface UserInfo {
+  readonly id: number;
+  readonly email: string;
+  readonly nickname: string;
+  readonly imageUrl?: string;
+  readonly refreshToken: string;
+}
