--- src/modules/auth/interfaces/userInfo.inerface.ts
@@ -1,7 +1,0 @@
-export interface UserInfo {
-  readonly id: number;
-  readonly email: string;
-  readonly nickname: string;
-  readonly imageUrl?: string;
-  readonly refreshToken: string;
-}
